move building map out
