+++ packages/dashboard/src/components/schedule-visualizer/tests/building-map-office.ts
